+++ src/pages/search-page/compare-terms/compare-terms.component.jsx
@@ -1,15 +1,16 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import localization from '../../../../lib/localization';
-import { getTranslateText } from '../../../../lib/translateText';
+import localization from '../../../lib/localization';
+import { getTranslateText } from '../../../lib/translateText';
 
 import './compare-terms.scss';
 
 export const CompareTerms = props => {
   const { uri, prefLabel, creator, onDeleteTerm } = props;
   let title = getTranslateText(prefLabel);
-  title = title.charAt(0).toUpperCase() + title.substring(1).toLowerCase();
+  title =
+    title && title.charAt(0).toUpperCase() + title.substring(1).toLowerCase();
   return (
     <div className="fdk-container p-4 fdk-container-compare-terms">
       <div className="d-flex align-items-baseline justify-content-between">
