+++ src/pages/search-page/compare-terms/compare-terms.scss
